pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skillos"
version = "0.1.0"
description = "Open-source implementation of SkillOS: Learning Skill Curation for Self-Evolving Agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "trl>=1.4.0",
    "transformers>=4.52.0",
    "peft>=0.15.0",
    "accelerate>=1.6.0",
    "vllm>=0.8.0",
    "torch>=2.6.0",
    "datasets>=3.0.0",
    "rank-bm25>=0.2.2",
    "pyyaml>=6.0",
    "alfworld>=0.3.0",
    "gymnasium>=0.29.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8.0",
]
reasoning = [
    "sympy>=1.13",
]
webshop = [
    "stable-baselines3>=2.6.0",
]

[tool.setuptools.packages.find]
include = ["skillos*"]

[tool.ruff]
line-length = 100
